+++ cdrs-tokio/src/cluster/connection_manager.rs
@@ -108,7 +108,10 @@ pub async fn startup<
                         .write_frame(&Frame::new_req_auth_response(response, version))
                         .await?;
                 }
-                ResponseBody::AuthSuccess(..) => break,
+                ResponseBody::AuthSuccess(success) => {
+                    authenticator.handle_success(success.data)?;
+                    break;
+                }
                 _ => return Err(format!("Unexpected auth response: {:?}", frame.opcode).into()),
             }
         }