+++ changelog.md
@@ -21,6 +21,7 @@ understated.
 * New `PreparedMetadataFlags`.
 * New `ClusterMetadata` representing information about a cluster.
 * Extracted protocol functionality to separate `cassandra-protocol` crate.
+* Passing final auth data from the server to `SaslAuthenticator`.
 
 ### Changed